Dec. 28, 2022 Update 1.0
UPDATE 1.0 
Added the files, tempy.py 

To Use: 
- Download the .py
- Put onto Raspi 
- Open with Geany, Thonny, etc, to build 

To Display Live: 
- Run via terminal 

To view log: 
- Open via desktop

Information: 
The log is currently checking stats every 1 second and saving into the log file. This is a simple, basic, test of my knowledge and application. Any confusion with multiple instances are resolved by a reboot. The log caps at 30 megabytes, enough for 24hrs + of logging and can be modified using the conversion tables you wish. 

This was made with assistance of RasPi Documentation, ChatGPT for error correction and formatting, and some common sense.